Module Generics

type 'a typ

The type for runtime representation of types.

type 'a dyn = 'a typ * 'a

The type for dynamic values, i.e., values with a type representation.

type any =

| Any : 'a typ -> any

Type representation value with existential variable.

Basic types

val unit : unit typ

Runtime representation for unit values.

val int : int typ

Runtime representation for integers.

val int32 : int32 typ

Runtime representation for the 32-bit integer type.

val int64 : int64 typ

Runtime representation for the 64-bit integer type.

val float : float typ

Runtime representation for the float type.

val bool : bool typ

Runtime representation for booleans.

val char : char typ

Runtime representation for characters.

val string : string typ

Runtime representation for strings.

val bytes : bytes typ

Runtime representation for bytes.

val option : 'a typ -> 'a option typ

Runtime representation for option types.

val result : 'a typ -> 'e typ -> ( 'a, 'e ) Stdlib.result typ

Runtime representation for result types.

Records

The following example demonstrates how to encode the runtime representation of a record:

type book = { title : string; authors : string list; pages : int }

let book_t =
  let open Generics in
  let title_field = field "title" string (fun b -> b.title) in
  let authors_field =
    field "authors" (list string) (fun b -> b.authors)
  in
  let pages_field = field "pages" int (fun b -> b.pages) in
  record "book" [ title_field; authors_field; pages_field ]
    (fun title authors pages -> { title; authors; pages })

module Field : sig ... end

Definitions for runtime field representation.

val field : string -> 'a typ -> ( 'record -> 'a ) -> ( 'record, 'a ) Field.t

field name field_t get : ('record, 'a) Field.t is the runtime representation of the field with type 'a, represented by field_t, belonging to the record 'record. See the Field module for more operations on fields.

module Record : sig ... end

Definitions for runtime record representation.

val record : string -> ( 'record, 'make ) Field.list -> 'make -> 'record typ

record name fields make is the runtime representation of the record type with a given name and fields. The function make can be used to create the record value from field values. See the Record module for more operations on records.

Variants

The following example demonstrates how to encode the runtime representation of the option type provided in the standard library.

(* type 'a option = None | Some of 'a *)

let option_t t =
  let open Generics in
  let none_constr = constr "None" (Const None) in
  let some_constr = constr "Some" (Args (t, fun x -> Some x)) in
  let value = function
    | None -> Constr.Value (none_constr, ())
    | Some x -> Constr.Value (some_constr, x)
  in
  variant "option"
    [ Constr.Any none_constr; Constr.Any some_constr ]
    value

module Constr : sig ... end

Definitions for runtime variant constructor representation. Constructors can either be constant or take one or more arguments.

val constr : string -> ( 'variant, 'args ) Constr.make -> ( 'variant, 'args ) Constr.t

constr name make : ('variant, 'args) Constr.t is the runtime representation of the constructor with arguments of type 'args belonging to the variant 'variant.

The make argument is either Const c, for a constant constructor c, or Args (args_t, variant_of_args) that represents the constructor with arguments fo type arg_t, where variant_of_args is the function from constructor arguments to the variant value.

See the Constr module for more operations on constructors.

module Variant : sig ... end

Definitions for runtime variant representation.

val variant : string -> 'variant Constr.any list -> ( 'variant -> 'variant Constr.value ) -> 'variant typ

Variant representation.

varinat name constr_list value is the runtime representation of a variant type with constructors described by constr_list where name is the type name of the variant and value is a function that maps every variant constructor to its runtime representation as a Constr.value.

Abstract types

val abstract : string -> 'a typ

Dynamic types

module Dyn : sig ... end

Definitions for values with runtime type representation.

Type equality

type ('a, 'b) equal =

| Equal : ( 'a, 'a ) equal

The type for equality between typ values.

val equal : 'a typ -> 'b typ -> ( 'a, 'b ) equal option

Checks for equality between two type representations.

Mapper

Mappers are datatype-generic interpreters. The most common examples of mappers provide functionality such as generic comparision or printing of values.

Custom mappers

The Map functor can be used to create user-defined generic mappers. The Mapper module type describes the required implementation of a mapper.

Every mapper needs to provide two types that represent the mapping and the conversion functions for each data type. The following example demonstrates this by implementing a simple mapper to convert values to strings:

module String_mapper = Generic.Map (struct
  (* The target mapping type from a value to a string. *)
  type 'a t = 'a -> string

  (* The type of the generic mapper. This is the type of the [self]
     argument passed to polymorphic mapping functions like [variant] and
     [record]. *)
  type mapper = { map : 'a. 'a typ -> 'a t }

  (* Mapping functions for basic types. *)
  let unit () = "()"
  let int = string_of_int
  let int32 = Int32.to_string
  let int64 = Int64.to_string
  let float = string_of_float
  let bool = string_of_bool
  let char = String.make 1
  let string x = String.concat "" [ "\""; String.escaped x; "\"" ]
  let bytes x = string (Bytes.to_string x)

  (* Mapping function for records. *)
  let record self record_t r1 =
    let fields =
      record_t
      |> Generics.Record.any_field_list
      |> List.map (fun (Generics.Field.Any field) ->
             let typ = Generics.Field.typ field in
             let show = self.map typ in
             let v = Generics.Field.get r1 field in
             String.concat ""
               [ "  "; Generics.Field.name field; " = "; show v; ";" ])
      |> String.concat "\n"
    in
    String.concat "\n" [ "{ "; fields; "}" ]

  (* Mapping function for variants. *)
  let variant self variant_t variant =
    let (Generics.Constr.Value (constr, args)) =
      Generics.Variant.value variant_t variant
    in
    let constr_name = Generics.Constr.name constr in
    match Generics.Constr.make constr with
    | Const _ -> constr_name
    | Args (args_t, _) ->
      let show_args = self.map args_t args in
      String.concat " " [ constr_name; "("; show_args; ")" ]

  (* Mapping function for abstract types. *)
  let abstract name _ = String.concat "" [ "<"; name; ">" ]
end)

The String_mapper module can now be used to convert any generic type to a string:

# String_mapper.map Generic.int 42;;
- : string = "42"

# String_mapper.map Generic.(option bool) (Some true);;
- : string = "Some (true)"

In addition to working on builtin types, String_mapper can be extended to work on abstract or user-defined types:

(* Define an abstract type and register a mapping function. *)
let out_channel_t : out_channel Generics.typ = Generics.abstract "out_channel" in
String_mapper.register out_channel_t
  (fun oc -> if oc = stdout then "<stdout>"
    else if oc = stderr then "<stderr>" else "<out_channel>");;

(* Call the generic mapper with our custom type. *)
# String_mapper.map out_channel_t stderr
- : string = "<stderr>"

module type Mapper = sig ... end

module Map (Mapper : Mapper) : sig ... end