Module Generics.Constr

Definitions for runtime variant constructor representation. Constructors can either be constant or take one or more arguments.

type ('variant, 'args) t

The runtime representation of constructors for the variant type 'variant with arguments of type 'args.

type ('variant, 'args) make =

| Const : 'variant -> ( 'variant, unit ) make

| Args : 'args typ * ( 'args -> 'variant ) -> ( 'variant, 'args ) make

The representation of the constructor values either as an immediate constant value or a function from constructor arguments to the variant value.

type 'variant value =

| Value : ( 'variant, 'args ) t * 'args -> 'variant value

The runtime representation of a constructor with arguments.

type 'variant any =

| Any : ( 'variant, 'args ) t -> 'variant any

The constructor with the hidden representation of the arguments.

val name : ( 'variant, 'args ) t -> string

The name of the constructor.

val make : ( 'variant, 'args ) t -> ( 'variant, 'args ) make

make constr is the representation of the value of the constructor.

The constructor can either be constant or have arguments.