Module Generics.Field

Definitions for runtime field representation.

type ('record, 'a) t

The runtime representation of fields with values of type 'a belonging to the record of type 'record.

type 'record any =

| Any : ( 'record, 'a ) t -> 'record any

type ('record, 'make) list =

| [] : ( 'record, 'record ) list

| :: : ( 'record, 'a ) t * ( 'record, 'make ) list -> ( 'record, 'a -> 'make ) list

The type for field lists that preserves individual field types.

The type variable 'make is the function type from record fields to the record type.

val name : ( 'record, 'a ) t -> string

name field is the name of field.

val typ : ( 'record, 'a ) t -> 'a typ

typ field is the runtime representation of the type of field.

val get : 'record -> ( 'record, 'a ) t -> 'a

get record field is the value of field in record.