Module Generics.Record

Definitions for runtime record representation.

type ('record, 'make) t

The runtime representation of records of type 'record.

The type variable 'make is the function type from field values to the record value. For example the record:

type r1 = { foo : int; bar : float list }

...has the type (r1, int -> float list -> r1) Record.t. This can be used to dynamically create record values given a list of fields (see Record.make).

val name : ( 'record, 'make ) t -> string

The name of the record type.

val any_field_list : ( 'record, 'make ) t -> 'record Field.any list

any_field_list record is the list of all record's fields represented as a list of Field.any to allow for uniform representation of individual fields.

val field_list : ( 'record, 'make ) t -> ( 'record, 'make ) Field.list

field_list record is the list of all record's fields represented as Field.list to preserve the types of individual fields.

val make : ( 'record, 'make ) t -> 'make

make record is a smart constructor function for the record represented as a function from fields to record of type 'record.